Refactor things out and add hot reloading of network mapped links.
